style(refactor): refactor to a run function
clear out main
