+++ src/cli/mod.rs
@@ -1,1 +1,2 @@
 pub mod commands;
+pub mod table;
